rename in_memory to relfect style guide
